# api/hqbase-mail-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HQBase Mail API",
    "version": "1.0.0",
    "description": "Stable v1 mail API for HQBase clients, automations, and agents. Additive fields may be introduced within v1; clients must ignore unknown response fields. Administrative APIs are not part of this contract."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://mail.example.com",
          "description": "HQBase origin"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/send": {
      "post": {
        "summary": "Send a new message",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sent message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Sending"],
        "operationId": "sendMessage",
        "description": "Not idempotent. Do not retry blindly.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInput"
              },
              "example": {
                "from": "support@example.com",
                "to": ["person@example.net"],
                "cc": [],
                "bcc": [],
                "subject": "Hello",
                "text": "Hello from HQBase",
                "attachmentIds": []
              }
            }
          }
        }
      }
    },
    "/api/v1/reply": {
      "post": {
        "summary": "Reply to a message",
        "security": [
          {
            "oauth2": ["mail:send"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sent reply",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Sending"],
        "operationId": "replyToMessage",
        "description": "Not idempotent. Do not retry blindly.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyInput"
              },
              "example": {
                "messageId": "msg_example",
                "from": "support@example.com",
                "cc": [],
                "bcc": [],
                "text": "Thanks for your message.",
                "attachmentIds": []
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Audience-bound OAuth token. Request resource={origin}/api/v1.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "refreshUrl": "/api/auth/oauth2/token",
            "scopes": {
              "mail:read": "Read accessible mail",
              "mail:write": "Organize accessible mail",
              "mail:send": "Manage drafts and send mail",
              "offline_access": "Request a refresh token"
            }
          }
        }
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Same-origin HQBase web session. Secure deployments may prefix the cookie name."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MessageSummary": {
        "type": "object",
        "required": [
          "id",
          "threadId",
          "mailboxId",
          "direction",
          "folder",
          "fromAddress",
          "to",
          "subject",
          "snippet",
          "receivedAt",
          "sentAt",
          "readAt",
          "starredAt",
          "hasAttachments",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "mailboxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "direction": {
            "type": "string",
            "enum": ["inbound", "outbound"]
          },
          "folder": {
            "type": "string",
            "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"]
          },
          "fromAddress": {
            "type": "string",
            "format": "email"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          },
          "receivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "sentAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "starredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "hasAttachments": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SendInput": {
        "type": "object",
        "required": ["from", "to", "subject", "text"],
        "properties": {
          "from": {
            "type": "string",
            "format": "email"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "maxItems": 50
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100000
          },
          "html": {
            "type": "string",
            "maxLength": 200000
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 20
          },
          "draftId": {
            "type": "string"
          }
        }
      },
      "ReplyInput": {
        "type": "object",
        "required": ["messageId", "from", "text"],
        "properties": {
          "messageId": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "format": "email"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "cc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "bcc": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 50
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100000
          },
          "html": {
            "type": "string",
            "maxLength": 200000
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 20
          },
          "draftId": {
            "type": "string"
          }
        }
      }
    }
  }
}